cloud_functions/requirements.txt
google-cloud-pubsub==2.10.0